fix(actions): remove invalid syntax
This change will remove the invalid sequence mentioned by GitHub. The workflow logic should be the same, just with a different syntax. The error message was:

Invalid workflow file

(Line: 5, Col: 17): A sequence was not expected.

Here is the link for reference: https://github.com/fidelity/homebrew-tap/actions/runs/16679528047/workflow

Signed-off-by: Casale, Robert <robert.casale@fmr.com>